fix true-color hilighting for file types that have marking char (fix #1769)

+++ far2l/src/console/interf.cpp
@@ -786,7 +786,7 @@ void ClearScreen(int Color)
 	Console.SetTextAttributes(Color);
 }
 
-int GetColor()
+DWORD64 GetColor()
 {
 	return (CurColor);
 }
